Remove accidental truncating conversion.
The intention of this code was to compute a mask for an `uint64_t`, but
accidentally performed to compuation with `unsigned long`. This
potentially lead to shifting beyond the width of the input value (e.g.,
for `range==63`).

We now enforce using wide enough types to remove this potential UB.

+++ hilti/runtime/include/types/integer.h
@@ -276,7 +276,7 @@ inline hilti::rt::integer::safe<UINT> bits(hilti::rt::integer::safe<UINT> v, uin
     if ( range == width )
         return v;
 
-    const auto mask = ((1U << range) - 1U) << lower;
+    const auto mask = ((uint64_t(1) << range) - uint64_t(1U)) << lower;
     return (v & mask) >> lower;
 }
